add stop stream button
