+++ train.py
@@ -1,7 +1,5 @@
 import os
 
-import llm_studio.app_utils.sections.chat_update
-
 os.environ["OMP_NUM_THREADS"] = "1"
 os.environ["MKL_NUM_THREADS"] = "1"
 os.environ["OPENBLAS_NUM_THREADS"] = "1"
@@ -561,10 +559,9 @@ def run_train_rlhf(
             with torch.no_grad():
                 logger.debug("Rollout: Generating response from active model")
                 output_dict = {}
-                output_dict["predicted_answer_ids"] = (
-                    llm_studio.app_utils.sections.chat_update.generate(batch, unwrap_model(model).cfg)
-                    .detach()
-                )
+                output_dict["predicted_answer_ids"] = model.generate(
+                    batch, unwrap_model(model).cfg
+                ).detach()
                 output_dict = train_dataloader.dataset.postprocess_batch_predictions(
                     output=output_dict
                 )
